Rewrote logic to use for loop since it's more performant for high numbers of geohashes. Added math utility class. Cleaned up code.
